Added PortMapper Function - Faster Discovery
Instead of port scanning on all subdomains, first we resolved it, uniquely sort it, and then do port scan and finally map open ports to associated subdomain. 
The function reads a list of resolved hosts from a file dnsxresolved.txt [Extracted from dnsx (dnsprobe.txt)] . 
For each domain, it searches a CSV file naabu.csv to extract the list of open ports. 
Then, it searches a text file dnsprobe.txt [output of dnsx] to identify all the subdomains associated with the domain. 
For each subdomain and port combination, it prints out a line in the format subdomain:port. 
The output is saved in a file hostport.txt.